+++ backstage/src/search-catalog.tsx
@@ -1,0 +1,8 @@
+import { BaseSearch } from "./components/base";
+import { withAuth } from "./lib/auth";
+
+function Command() {
+  return <BaseSearch key="software-catalog" type="software-catalog" />;
+}
+
+export default withAuth(Command);