+++ backstage/src/search-techdocs.tsx
@@ -1,0 +1,8 @@
+import { BaseSearch } from "./components/base";
+import { withAuth } from "./lib/auth";
+
+function Command() {
+  return <BaseSearch key="techdocs" type="techdocs" />;
+}
+
+export default withAuth(Command);