+++ backstage/src/search.tsx
@@ -1,0 +1,8 @@
+import { BaseSearch } from "./components/base";
+import { withAuth } from "./lib/auth";
+
+function Command() {
+  return <BaseSearch key="all" type="all" />;
+}
+
+export default withAuth(Command);